Code executes program 1, not O-scope verified
