+++ Project_6/source/app.h
@@ -3,5 +3,6 @@
 
 
 
-
+/* T A S K S */
 void prv_GenerateSineWave(void *prvParameters);
+void prv_TimerTask(void *prvParameters);
