+++ Project_6/source/main.c
@@ -44,13 +44,16 @@
 #include "gpio.h"
 #include "Systick.h"
 #include "myTimer.h"
-
+#include "app.h"
 
 /* F R E E   R T O S   I N C L U D E S */
 #include "FreeRTOS.h"
 #include "task.h"
 #include "queue.h"
 #include "timers.h"
+
+
+static bool lightOn;
 
 /* D E F I N I T I O N S */
 
@@ -62,6 +65,24 @@ static void hello_task(void *pvParameters);
 /*
  * @brief   Application entry point.
  */
+
+void myTimerCallback( TimerHandle_t timer)
+{
+	DacIncrementAndSet();
+	if(!lightOn)
+	{
+		gpioGreenLEDOn();
+		lightOn = true;
+	}
+	else
+	{
+		gpioLEDsOff();
+		lightOn = false;
+	}
+
+}
+
+
 int main(void) {
 
   	/* Init board hardware. */
@@ -74,9 +95,25 @@ int main(void) {
     /* Initialize application modules */
     logInit(LL_Debug);
     gpioInit();
+    DacInit();
+
+    /* Create timer module */
+    lightOn = false;
+    TickType_t timerPeriod = pdMS_TO_TICKS(100);
+    TimerHandle_t p1Timer =  xTimerCreate(".1s timer",
+    							timerPeriod,
+    							pdTRUE,
+    							(void *) 2,
+    							myTimerCallback);
+
+    xTimerStart(p1Timer, 0);		// Start timer, wait 0 ticks
+
 
     /* Create Tasks */
-    xTaskCreate(prv_GenerateSineWave, "Generate_Sine_Wave", configMINIMAL_STACK_SIZE + 10, NULL, genwave_PRIORITY, NULL);
+
+    BaseType_t timer_task_handle;
+//    timer_task_handle = xTaskCreate(prv_TimerTask, "Timer Task", configTIMER_TASK_STACK_DEPTH, NULL, configTIMER_TASK_PRIORITY, NULL);
+//    xTaskCreate(prv_GenerateSineWave, "Generate_Sine_Wave", configMINIMAL_STACK_SIZE + 10, NULL, genwave_PRIORITY, NULL);
 //    xTaskCreate(hello_task, "Hello_task", configMINIMAL_STACK_SIZE + 10, NULL, hello_task_PRIORITY, NULL);
     vTaskStartScheduler();
     while(1);
